Formatting

+++ src/main/kotlin/com/beust/sixty/Extensions.kt
@@ -2,7 +2,7 @@ package com.beust.sixty
 
 import javafx.scene.paint.Color
 
-fun Byte.h(): String = String.format("%02x", this.toInt()).toUpperCase()
+fun Byte.h(): String = String.format("%02x", this).toUpperCase()
 fun Int.h(): String = String.format("%02x", this).toUpperCase()
 fun Int.hh(): String = String.format("%04x", this).toUpperCase()
